@@ -1,7 +1,4 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { Subscription } from 'rxjs';
-import { environment } from 'src/environments/environment';
-import { PetListItem } from '../pet-list-item';
+import { Component } from '@angular/core';
 import { PetService } from '../pet.service';
 
 @Component({
@@ -9,19 +6,11 @@ import { PetService } from '../pet.service';
   templateUrl: './account-pets.component.html',
   styleUrls: ['./account-pets.component.css']
 })
-export class AccountPetsComponent implements OnInit, OnDestroy{
+export class AccountPetsComponent{
 
-  apiUrl = environment.apiUrl + '/';
-  pets!: PetListItem[];
-  sub!: Subscription;
+  pets$ = this.petService.apiRequestListItem$;
 
-  constructor(private petService: PetService) { }
-
-  ngOnInit() {
-    this.sub = this.petService.pets$.subscribe(pets => this.pets = pets);
+  constructor(private petService: PetService) {
   }
 
-  ngOnDestroy() {
-    this.sub.unsubscribe();
-  }
 }
