@@ -4,5 +4,7 @@ export const environment = {
   petfinderApiSecret: '',
   petfinderVersion: 2,
   petfinderBaseUrl: 'https://api.petfinder.com',
-  apiUrl: 'https://app-pawssierapi.azurewebsites.net'
+  apiUrl: 'https://app-pawssierapi.azurewebsites.net',
+  graphUrl: 'https://graph.microsoft.com/v1.0/me',
+  postLogoutUrl: 'http://pawssier.azurewebsites.net/homepage'
 };
